Move valid_address to validators
