+++ emails/tests/validator_tests.py
@@ -5,10 +5,24 @@ from unittest.mock import Mock, patch
 from django.contrib.auth.models import User
 from django.test import TestCase
 
+from waffle.testutils import override_flag
+
 from ..exceptions import CannotMakeSubdomainException
-from ..models import Profile, RegisteredSubdomain, hash_subdomain
-from ..validators import has_bad_words, is_blocklisted, valid_available_subdomain
-from .models_tests import make_premium_test_user
+from ..models import (
+    DomainAddress,
+    Profile,
+    RegisteredSubdomain,
+    RelayAddress,
+    hash_subdomain,
+)
+from ..validators import (
+    has_bad_words,
+    is_blocklisted,
+    valid_address,
+    valid_address_pattern,
+    valid_available_subdomain,
+)
+from .models_tests import make_free_test_user, make_premium_test_user
 
 
 class HasBadWordsTest(TestCase):
@@ -126,3 +140,44 @@ class ValidAvailableSubdomainTest(TestCase):
     def test_subdomain_with_space_at_end_raises(self) -> None:
         with self.assertRaisesMessage(CannotMakeSubdomainException, self.ERR_NOT_AVAIL):
             valid_available_subdomain("mydomain ")
+
+
+class ValidAddressPatternTest(TestCase):
+    def test_valid_address_pattern_is_valid(self) -> None:
+        assert valid_address_pattern("foo")
+        assert valid_address_pattern("foo-bar")
+        assert valid_address_pattern("foo.bar")
+        assert valid_address_pattern("f00bar")
+        assert valid_address_pattern("123foo")
+        assert valid_address_pattern("123")
+
+    def test_valid_address_pattern_is_not_valid(self) -> None:
+        assert not valid_address_pattern("-")
+        assert not valid_address_pattern("-foo")
+        assert not valid_address_pattern("foo-")
+        assert not valid_address_pattern(".foo")
+        assert not valid_address_pattern("foo.")
+        assert not valid_address_pattern("foo bar")
+        assert not valid_address_pattern("Foo")
+
+
+class ValidAddressTest(TestCase):
+    def test_valid_address_dupe_of_deleted_invalid(self) -> None:
+        user = make_free_test_user()
+        relay_address = RelayAddress.objects.create(user=user)
+        relay_address.delete()
+        assert not valid_address(relay_address.address, relay_address.domain_value)
+
+    @override_flag("custom_domain_management_redesign", active=True)
+    def test_valid_address_dupe_domain_address_of_deleted_is_not_valid(self) -> None:
+        user = make_premium_test_user()
+        user.profile.subdomain = "mysubdomain"
+        user.profile.save()
+        address = "same-address"
+        domain_address = DomainAddress.make_domain_address(
+            user.profile, address=address
+        )
+        domain_address.delete()
+        assert not valid_address(
+            address, domain_address.domain_value, user.profile.subdomain
+        )
